scripts/agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-agent"
version = "0.1.0"
description = "Autonomous agent framework for enterprise workflows"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.0.50",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.1",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agents"]
